Move referencedat widget to global property #22069
