+++ projects/safe/src/lib/services/form-builder.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
 import * as Survey from 'survey-angular';
-import { renderCustomProperties } from '../survey/custom-properties';
+import { SafeReferenceDataService } from './reference-data.service';
+import { renderGlobalProperties } from '../survey/render-global-properties';
 
 /**
  * Shared form builder service.
@@ -13,10 +14,9 @@ export class SafeFormBuilderService {
   /**
    * Constructor of the service
    *
-   * @param domService The dom service
    * @param referenceDataService Reference data service
    */
-  constructor() {}
+  constructor(private referenceDataService: SafeReferenceDataService) {}
 
   /**
    * Creates new survey from the structure and add on complete expression to it.
@@ -26,7 +26,9 @@ export class SafeFormBuilderService {
    */
   createSurvey(structure: string): Survey.Survey {
     const survey = new Survey.Model(structure);
-    survey.onAfterRenderQuestion.add(renderCustomProperties());
+    survey.onAfterRenderQuestion.add(
+      renderGlobalProperties(this.referenceDataService)
+    );
     const onCompleteExpression = survey.toJSON().onCompleteExpression;
     if (onCompleteExpression) {
       survey.onCompleting.add(() => {
