+++ projects/safe/src/lib/survey/creator.ts
@@ -1,19 +1,20 @@
+import { JsonMetadata, Question } from 'survey-angular';
+
 /**
  * Edits general settings of the survey builder.
  *
  * @param Survey Survey library
  */
 export const initCreatorSettings = (Survey: any): void => {
-  Survey.Serializer.findProperty('question', 'name').readOnly = true;
-  Survey.Serializer.findProperty('question', 'name').dependsOn = 'valueName';
-  Survey.Serializer.findProperty('question', 'name').onGetValue = (obj: any) =>
+  const serializer: JsonMetadata = Survey.Serializer;
+
+  serializer.findProperty('question', 'name').readOnly = true;
+  serializer.findProperty('question', 'name').onGetValue = (obj: Question) =>
     obj.valueName ? obj.valueName : obj.name;
-  Survey.Serializer.findProperty('question', 'valueName').isRequired = true;
-  Survey.Serializer.findProperty('file', 'storeDataAsText').onGetValue = (
-    obj: any
-  ) => false;
-  Survey.Serializer.findProperty('file', 'storeDataAsText').readOnly = true;
-  Survey.Serializer.findProperty('file', 'storeDataAsText').visible = false;
-  Survey.Serializer.findProperty('file', 'maxSize').onGetValue = (obj: any) =>
-    7340032;
+  serializer.findProperty('question', 'valueName').isRequired = true;
+  serializer.findProperty('file', 'storeDataAsText').onGetValue = (obj: any) =>
+    false;
+  serializer.findProperty('file', 'storeDataAsText').readOnly = true;
+  serializer.findProperty('file', 'storeDataAsText').visible = false;
+  serializer.findProperty('file', 'maxSize').onGetValue = (obj: any) => 7340032;
 };
